Clean up code & add content to readme
