+++ .gitignore
@@ -2,3 +2,4 @@
 # Ignore dynaconf secret files
 .secrets.*
 settings.toml
+__pycache__
